requirements.txt
langchain
langgraph
langgraph-cli[inmem]
langchain-google-genai
langchain-google-community[gmail,sheets]
google-auth-oauthlib
google-api-python-client
python-dotenv
langsmith
deepagents
python-docx
duckduckgo-search
python-jobspy
playwright
gspread
pytest
fastapi
uvicorn[standard]
aiohttp
aiosqlite
jinja2
google-auth
